feat!: B6e Phase 1 — decimal SSOT + all-text Excel across all four bindings (float principle) (#120)
Every binding parses decimals through the verified Agda kernel SSOT (aletheia_parse_decimal); the per-binding float→rational heuristics are deleted. Numeric API params take exact types (Python int | Fraction; Go Rational); float AND bool are rejected at the exact-numeric boundary. Excel loaders adopt an all-text contract (a number cell is rejected with row+column+value+reason). Decimal parsing is RTS-gated and vocal. Docs/demos updated to the float principle; a demo gate runs every example end-to-end.

BREAKING: per-binding float→rational APIs removed; numeric params are exact types; Excel numeric cells must be text; YAML/Excel loaders require a live backend.

+++ python/aletheia/asyncio/_client.py
@@ -327,7 +327,7 @@ class AletheiaClient:  # pylint: disable=too-many-public-methods
         can_id: int,
         dlc: DLCCode,
         frame: bytes | bytearray,
-        signals: Mapping[str, float | Fraction],
+        signals: Mapping[str, int | Fraction],
         *,
         extended: bool = False,
     ) -> bytearray:
@@ -345,7 +345,7 @@ class AletheiaClient:  # pylint: disable=too-many-public-methods
         self,
         can_id: int,
         dlc: DLCCode,
-        signals: Mapping[str, float | Fraction],
+        signals: Mapping[str, int | Fraction],
         *,
         extended: bool = False,
     ) -> bytearray:
